Clean up mistaken use of references as layers list
Original code was walking manifest.References() which includes more than
just the layers. Then the code was (wrongly) adding the reference to the
config (manifest-v2 JSON); so for a single layer image, the `inspect`
was reporting 3 layers (2 instances of config, and 1 = the real layer);
but was also causing an extra cross-repo mount of the config blob.

Signed-off-by: Phil Estes <estesp@gmail.com>