Better handling of v2-only registry interaction
Get rid of v1 endpoint ping (unecessary, as the entire support for
manifest lists requires a v2 endpoint anyway) and gracefully handle
potential registry limitations on getting all tag list. The IBM Bluemix
registry currently blocks the "list all tags" API endpoint, but it is
not required for inspect, as it just limits the output data in a small
way.

Signed-off-by: Phil Estes <estesp@gmail.com>