Dynamically generate the os/arch list
Switch to validate OS and arch separately for a more direct alignment
with the OCI spec allowance for platform definition. This means you can
easily create platform objects that may not make sense, but fixes the
issue of adversely restricting creation of platform objects that aren't
recognized yet in the list of Go-compilable platforms.

Signed-off-by: Phil Estes <estesp@gmail.com>